Moved the networks into its own subdirectory.
